Fix OSX build - no idea why it wasn't breaking before (nw)

+++ src/osd/modules/debugger/osx/debugconsole.mm
@@ -160,7 +160,7 @@
 
 	windowFrame.size.width += adjustment.width;
 	windowFrame.size.height += adjustment.height; // not used - better to go for fixed height
-	windowFrame.size.height = std::min(512.0, available.size.height);
+	windowFrame.size.height = std::min(CGFloat(512.0), available.size.height);
 	windowFrame.size.width = std::min(windowFrame.size.width, available.size.width);
 	windowFrame.origin.x = available.origin.x + available.size.width - windowFrame.size.width;
 	windowFrame.origin.y = available.origin.y;
